+++ exercises/latex/05_c/listings/accelerometerLCDFormat.c
@@ -1,0 +1,6 @@
+//*** Beschleunigungssensor ***
+//Orientierung X:
+//Orientierung Y:
+//Orientierung Z:
+
+//Die Ebene ist (nicht) waagrecht.