+++ exercises/templates/WinIdeaWorkspaceTemplate/lib/gfx.c
@@ -2,14 +2,15 @@
  *  This code is based on:  https://github.com/adafruit/Adafruit-GFX-Library &  https://github.com/adafruit/TFTLCD-Library
  *  The code is converted by Puria Izady for the Cypress FM4  microcontroller
  */
+#include <stdlib.h>
+
 #include "gfx.h"
 #include "lcd.h"
 #include "src/display.h"
-#include <stdlib.h>
 
 static const char *WHITESPACE = " ";
 
-// Fix for the missing declaration of ito in stdlib.h.
+// Fix for the missing declaration of itoa in stdlib.h.
 char* itoa(int, char* , int);
 
 /**
@@ -554,4 +555,35 @@ void write16BitDigit(const uint16_t *value, uint8_t mode){
     itoa(*value, buffer, base);
     writeText_s(buffer);
   }
+}
+
+void write8BitDigit(const uint8_t *value){
+  const uint8_t base = 10;
+    uint8_t numberOfDigits = 0;
+    if (*value < 1000) {
+      numberOfDigits = 3;
+      if (*value < 100) {
+        numberOfDigits = 2;
+        if (*value < 10) {
+         numberOfDigits = 1;
+        }
+      }
+    }
+    char buffer[numberOfDigits];
+    itoa(*value, buffer, base);
+    writeText_s(buffer);  
+}
+
+void writeFloat(float number, uint8_t precision, uint8_t width){
+  char buffer[width];
+  
+  char settings[4];
+  strcpy(settings, "%.");
+  char widthBuffer[1];
+  itoa(precision, widthBuffer, 10);
+  strcat(settings, widthBuffer);
+  strcat(settings, "f");
+
+  int test = sprintf(buffer, settings, number);
+  writeText_s(buffer);
 }